Add plugin providing task to release staged repositories
When publishing to mvn central repo (either snapshot or release) we
still push artifacts to staging environement. In order for those to be
accessible publicly we need to release those. One option is to do it
manually via the web ui but this is not too convinient and fortunately
its possible to use the plugin to introduce new gradle tasks handling
that.

more to read about here:
- https://central.sonatype.org/pages/gradle.html
- https://github.com/Codearte/gradle-nexus-staging-plugin/